requirements.txt
azure-ai-inference>=1.0.0
requests>=2.28.0
PyYAML>=6.0
